Fix bug; add Prop cost model, extractor, and tests
Fix mistake in Prop logic grammar where only terms and not props
could be negated.

Implement cost model and term extractor for Prop grammar (same idea
as for exprs, but with different weights) and add tests.

Currently, there is an absurd overlap between the cost models and
term extractors for Props and Exprs. TODO: abstract the
int-based cost computation and minimization logic so it's more
easily reusable.