Update expr tests with separate ExprTree class
This means that we test both of the expected client use cases:
(1) Prop tests when the user has access the tree class and can
make it extend `QuicheTree` and add `value()` and `children()`
methods to it.
(2) Expr tests when the user creates a wrapper class that extends
`QuicheTree` for a tree class they can't/don't want to modify.