Intrusive update/bugfix/refactor
The goal was to fix issue where contrapositive test was not
correctly inferring (-> x y) from (-> ~y ~x). It turned out
that the problem was that e-matching only applied the first
match of a rule in each eclass, and if we have commutative
operators, we need to be able to apply rewrites to all enodes
in each eclass.

This means that internal ematch functions now return a list of
environments specifying all ways a rule can be applied.

This fixes the issue with inferring contrapositives and more
generally with allowing rewrite rules for commutativity.

However, there are definitely inefficiencies in the ematch
implementation, and a review for performance improvements
may be needed before this can scale to full program ASTs.

Other changes:
* add `QuicheTree.is_pattern_symbol()` to support
user-specified "pattern symbols" to
distinguish symbols from patterns in rewrite rules
* update test suite so that rewrite rules use pattern symbols
* fix test function `verify_egraph_shape()` to use tuples
instead of lists to correctly verify eclasses with multiple
enodes using the same key (i.e., support commutativity rules)
* update Prop lexer to support ? as a pattern symbol