Add conditional rules + Python arithmetic rewrites
* Add ConditionalRule class that supports conditional rewrite rules by
either passing in a checker function to __init__ or by extending the
ConditionalRule class and implementing `check_condition`.

* Add quiche.pyast.pyarith_rewrites with lots of simple arithmetic
rewrite rules (e.g., x*1 = x, x/x = 1 if x != 0 (example of
conditional rule), and lots more). Use `get_all_arith_rules()` for a
list of all of the Rule objects defined in the file. May in theory be
passed to `Rule.apply_rules(rules, egraph)` but in practice this causes
an overload because we don't have a scheduler or cutoff options (guess
what's coming soon!).

* Add a new file to start testing arithmetic rewrites (incomplete)